reroutes fixed
fixed a bug where, if trying to manually enter a link to customer or admin page without being logged in, the user would be sent to the my-account login page, but nothing would be active. had to $this->_f3->reroute('my-account'); instead of what i was doing before (was rendering my-account page with Template instead, which didn't load most of the code allowing login verification and such).